Cleanup of source files
