+++ src/Apps/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB SPINE_SOURCES_APPS ${CMAKE_CURRENT_SOURCE_DIR}/*.cc)
+set(SPINE_SOURCES_APPS ${SPINE_SOURCES_APPS} PARENT_SCOPE)
+set(SPINE_APP_DIR ${CMAKE_CURRENT_SOURCE_DIR} PARENT_SCOPE)
