+++ src/Base/CMakeLists.txt
@@ -1,0 +1,2 @@
+file(GLOB SPINE_SOURCES_BASE ${CMAKE_CURRENT_SOURCE_DIR}/*.cc)
+set(SPINE_SOURCES_BASE ${SPINE_SOURCES_BASE} PARENT_SCOPE)
