+++ src/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_subdirectory(Apps)
+add_subdirectory(Base)
+
+set(SPINE_SOURCES ${SPINE_SOURCES_APPS} ${SPINE_SOURCES_BASE} PARENT_SCOPE)
+set(SPINE_SOURCES_APPS ${SPINE_SOURCES_APPS} PARENT_SCOPE)
+set(SPINE_SOURCES_BASE ${SPINE_SOURCES_BASE} PARENT_SCOPE)
+set(SPINE_APP_DIR ${SPINE_APP_DIR} PARENT_SCOPE)
